+++ src/main/java/com/spendingTracker/spendingItem/SpendingItemController.java
@@ -22,7 +22,7 @@ public class SpendingItemController {
         List<SpendingItem> expenses = repository.findAll();
         List<SpendingItem> returnExpenses = new ArrayList<SpendingItem>();
         for (int i = 0; i < expenses.size(); i++){
-            if (expenses.get(i).getUserId() == Float.parseFloat(userId)){
+            if (expenses.get(i).getUserId() == Long.parseLong(userId)){
                 returnExpenses.add(expenses.get(i));
             }
         }
